Create Challenge
Knowledge Management in Projects - AI

The dynamics and requirements in project management are rapidly increasing, presenting a challenge for all involved parties. Resources and knowledge are fundamentally available, but not constantly (24/7) – precisely when they are needed. Often, it is unclear how inquiries need to be formulated to receive correct answers. Given the multitude of sources (data) and resources (internal as well as external stakeholders), the goal is to provide a continuous, AI-supported service using technology. This involves integrating sources such as SIA standards and external "Open Data" sources to support project-related queries contextually and in real-time.